changed storage path to be based on env

+++ mycloud-2.0/app/api/files/delete/route.ts
@@ -15,6 +15,13 @@ export const DELETE = async (req: NextRequest) => {
         );
     }
 
+    if (!process.env.FILE_STORAGE_PATH) {
+        return NextResponse.json(
+        { error: "Storage path not set" },
+        { status: 400 }
+        );
+    }
+
     const {id} = await req.json();
 
     const file = await prisma.file.delete({
@@ -28,7 +35,7 @@ export const DELETE = async (req: NextRequest) => {
         }
     });
 
-    const filePath = path.join(process.cwd(), "public", "test_storage", session.user.id.toString(), file.name);
+    const filePath = path.join(process.env.FILE_STORAGE_PATH, session.user.id.toString(), file.name);
     await unlink(filePath);
 
     return NextResponse.json({ message: "File removed" });
